requirements.txt
Django==1.10
PyJWT==1.4.2
Pygments==2.1.3
argparse==1.2.1
backports.shutil-get-terminal-size==1.0.0
decorator==4.0.10
djangorestframework==3.4.6
djangorestframework-jwt==1.8.0
ipdb==0.10.1
ipython==5.1.0
ipython-genutils==0.1.0
pathlib2==2.1.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.7
ptyprocess==0.5.1
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.2
wcwidth==0.1.7
wsgiref==0.1.2
